Clean up <* 🎉 2020
